Make pom dependency optional (but default)
