+++ src/reader.rs
@@ -1,3 +1,5 @@
+#![cfg(any(feature = "pom_parser", feature = "nom_parser"))]
+
 use log::{error, warn};
 use std::cmp;
 use std::convert::TryInto;
